refactor: extract DumpService interface with NoOpDumpService for non-local profiles
DumpService was a concrete @Profile("local") class, leaving HtmlCleaner
without a required bean on other profiles. Extracted an interface, moved
the local implementation to LocalDumpService, and added a no-op
implementation activated via @Profile("!local").
