+++ extractor/src/main/java/net/shamansoft/cookbook/html/HtmlCleaner.java
@@ -1,13 +1,16 @@
 package net.shamansoft.cookbook.html;
 
+import com.google.type.DateTime;
 import io.micrometer.core.instrument.MeterRegistry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.shamansoft.cookbook.config.HtmlCleanupConfig;
 import net.shamansoft.cookbook.html.strategy.CleanupStrategy;
 import net.shamansoft.cookbook.html.strategy.Strategy;
+import net.shamansoft.cookbook.service.DumpService;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.Optional;
 
 /**
@@ -26,6 +29,7 @@ public class HtmlCleaner {
     private final HtmlCleanupConfig config;
     private final MeterRegistry meterRegistry;
     private final java.util.List<CleanupStrategy> strategies;
+    private final DumpService dumpService;
 
     /**
      * Preprocess HTML using the hybrid strategy cascade.
@@ -43,6 +47,8 @@ public class HtmlCleaner {
             log.warn("Empty HTML input for URL: {}", url);
             return buildResult("", 0, Strategy.FALLBACK);
         }
+
+        dumpService.dump(html, "cleanerInput", "html", Instant.now().toString());
 
         if (!config.isEnabled()) {
             return buildResult(html, originalSize, Strategy.DISABLED);
@@ -92,7 +98,7 @@ public class HtmlCleaner {
         // Emit metrics
         emitMetrics(strategy, originalSize, cleanedSize, reductionRatio);
 
-        return new Results(
+        Results results = new Results(
                 cleanedHtml,
                 originalSize,
                 cleanedSize,
@@ -100,6 +106,9 @@ public class HtmlCleaner {
                 strategy,
                 message
         );
+
+        dumpService.dump(results.toString(), "cleanerOutput-"+results.strategyUsed(), "txt", Instant.now().toString());
+        return results;
     }
 
     /**
